updates revaulting flow
